+++ CHANGELOG.md
@@ -1,7 +1,19 @@
 CHANGELOG
 ---------
 
+- **March.3.2016**: 1.0.5 - Breaking Changes
+  - Move spreadsheet_columns method from the class to the instance
+  - Fix Bug: remove default underline on cells
+- **March.1.2016**: 1.0.4
+  - Extract helper methods to seperate module
+  - Improve readme
+- **March.1.2016**: 1.0.3
+  - Fix/Improve renderers
+  - Fix header default background color
+  - Fix default columns
+- **February.26.2016**: 1.0.2
+  - Enhance Style options
 - **February.26.2016**: 1.0.1
-  - Fix Bug in ActionController Renderers
+  - Fix bug in renderers
 - **February.26.2016**: 1.0.0
   - Gem Initial Release